+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - remove transaction for catalog_product_flat reindexAll
 - add config to toggle the transaction for catalog_product_flat
 - add configurable cron for reindex
+- add config to toggle the transaction for catalog_product_price
+